Improve responsivity, Capacitive touch working nicely
Turns out that the `delayMicroseconds()` function does not play nicely on Teensy4.0, changing the delay function to just `delay()` appears to work well with the Ultrasonic Sensor.

With this, the foundation of the Paddle_of_Theseus is pretty solid; the operations of "fret", "strum", and "whammy" work reliably and create good sounds on Ableton.

Still TODO:
* Get the Rotary Potentiometer for static wet/dry effects or something
* Now purchase several Cap Touch sensors and use them for chords
* Document HW connections via Fritzing, Eagle, or something
* Integrate Accelerometer to modify the sound
* Mount these sensors onto some kind of first-take instrument.