Rotary Encoder Master Volume support
Again just a quick hack to put master volume into one of the rotary encoder settings. This one is indicated when the RGB LED is completely off on the Rotary Encoder. I still want to fully clean up this code, likely by turning the capacitive touch sensor code into 3 instances of a class. But in the meantime, I can understand this *and* it works.