Major Graphics and layout improvements
Change everything from white rectangles to something a bit more stylized. Still much work to be done in terms of polish, but these first changes have helped to tease out what is possible and the whole theme is probably closer to looking good than it seems. Certainly it will be good to go back and clean up the style later when things get a bit more solid.