Implement heartbeat thread initial commit
Resolves #13
Resolves #15
Create a heartbeat thread when the paddle receives a serial connection, send a message every 5 seconds and return to "waiting for serial connection" splash screen when the connection is lost. This initial commit is validated on the paddle and gracefully handles unplug and re-plug events.

Also substitute print for thread-safe "logging" lib and improve the splash screen to override mouseClickEvent